Refactored add list of relations logic
